+++ src/Animal.h
@@ -23,7 +23,7 @@ class Animal {
 		// untuk mengecek liar apa tidaknya
 
 		//virtual void Interract() = 0;
-		static int GetNbMakanan();
+		static float GetNbMakanan();
 
 	protected:
 //		enum eatingType {
@@ -34,7 +34,7 @@ class Animal {
 		const int BeratBadan;
 		//const bool amfibi; //amfibi/tidak
 		//lokasi cell
-		static int NbMakanan;
+		static float NbMakanan;
 		//lokasi cel--> friend class sama cage aja, ntr dipanggil di cage?
 
 };
